Created Board class
 - Renamed moves.py to components.py
 - Renamed boards.py to test_boards.py
 - components.py
    - Added board class
    - Handles board as an object
    - Renamed get_moves to find_all_moves and added as Board function
 - benchmarks.py
    - Updated to use new Board class
